+++ examples/search/seq_r0_v1.py
@@ -2,6 +2,7 @@ import numpy as np
 from scipy.optimize import minimize
 import subprocess
 import sys
+import os
 #----------------------------------------------------------------------
 # command: python3 nm_r0_v1.py
 #----------------------------------------------------------------------
@@ -25,19 +26,6 @@ subprocess.run("echo \"No.: ETA value [eV], r0, sigma\" > Evalute.txt", shell=Tr
 #----------------------------------------------------------------------
 # fitting parameters
 element = "B"
-#------------------------
-#rcov = 0.84      # Covalentradii [Angstrom]
-#r0  = rcov/0.529*1.85 # Radius [bohr] = [Angstrom]/0.529
-#sigma = 2.0      # Power
-#------------------------
-#area=[
-#  (r0,3.0*r0),
-#  (2.0,14.0)
-#]
-#------------------------
-#print("initial parameters, r0: "+str(r0))
-#print("initial parameters, sigma: "+str(sigma))
-#x0 = np.array([r0,sigma])
 
 subprocess.run("cd ./"+str(element)+" ; rm -f -r results ; cd ../", shell=True)
 subprocess.run("cd ./"+str(element)+" ; mkdir results ; cd ../", shell=True)
@@ -56,21 +44,22 @@ def f(r0,sigma):
   subprocess.call("sed -i s/r0/"+str(r0)+"/g "+str(file_inp), shell=True)
   subprocess.call("sed -i s/sigma/"+str(sigma)+"/g "+str(file_inp), shell=True)
   
-  sub = subprocess.run("/mnt/d/skprogs/sktools/src/sktools/scripts/skgen.py -o slateratom -t sktwocnt sktable -d "+str(element)+" "+str(element), shell=True)
+  skgen = subprocess.run("/mnt/d/skprogs/sktools/src/sktools/scripts/skgen.py -o slateratom -t sktwocnt sktable -d "+str(element)+" "+str(element), shell=True, stdout=subprocess.PIPE)
   
-  subprocess.run("cd ./"+str(element)+" ; ./run.sh ; cd ../", shell=True)
-  
-  evaluate = subprocess.run("awk '{if(NR==10){printf \"%s\",$3}}' ./"+str(element)+"/"+str(file_msd), shell=True, stdout=subprocess.PIPE)
-  if evaluate.returncode == 0:
-    y = float(str(evaluate.stdout).lstrip("b'").rstrip("\\n'"))
+  if os.path.exists(str(element)+"-"+str(element)+".skf"): 
+    evaluate = subprocess.run("awk '{if(NR==10){printf \"%s\",$3}}' ./"+str(element)+"/"+str(file_msd), shell=True, stdout=subprocess.PIPE)
+    if evaluate.returncode == 0:
+      y = float(str(evaluate.stdout).lstrip("b'").rstrip("\\n'"))
+      subprocess.run("mv "+str(file_inp)+" ./"+str(element)+"/results/"+str(file_inp)+"_No"+str(count), shell=True)
+      subprocess.run("cp ./"+str(element)+"/comp_band.png ./"+str(element)+"/results/comp_band_No"+str(count)+".png", shell=True)
+    else:
+      y = 99999.999
   else:
     y = 99999.999
   
   print("Evaluate: ",str(y))
   print("Parameters: "+"[ "+str(r0)+","+str(sigma)+" ]")
   print("------------------------")
-  subprocess.run("mv "+str(file_inp)+" ./"+str(element)+"/results/"+str(file_inp)+"_No"+str(count), shell=True)
-  subprocess.run("cp ./"+str(element)+"/comp_band.png ./"+str(element)+"/results/comp_band_No"+str(count)+".png", shell=True)
   subprocess.run("echo No."+str(count)+": "+str(y)+", "+str(r0)+", "+str(sigma)+" >> Evalute.txt", shell=True)
 
   return y
